requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
jinja2
httpx
python-multipart
python-dotenv
websockets
bcrypts
PyJWT